fixed and added unittests

+++ pyjob/platform/tests/test_lsf.py
@@ -106,6 +106,28 @@ class TestLoadSharingFacility(unittest.TestCase):
         for f in jobs:
             os.unlink(f)
 
+    def test_sub_2(self):
+        assert "PYJOB_ENV" not in os.environ
+        os.environ["PYJOB_ENV"] = "pyjob_random"
+        jobs = [make_script(["echo $PYJOB_ENV"])]
+        jobid = LoadSharingFacility.sub(jobs, name=inspect.stack()[0][3])
+        time.sleep(5)
+        start, timeout = time.time(), False
+        while LoadSharingFacility.stat(jobid):
+            # Don't wait too long, one minute, then fail
+            if ((time.time() - start) // 60) >= 1:
+                LoadSharingFacility.kill(jobid)
+                timeout = True
+            time.sleep(10)
+        for f in jobs:
+            os.unlink(f)
+        if timeout:
+            self.assertEqual(1, 0, "Timeout")
+        else:
+            log = jobs[0].replace(".sh", ".log")
+            self.assertTrue(os.path.isfile(log))
+            self.assertTrue(open(log).read().strip(), "pyjob_random")
+
     # ================================================================================
     # ARRAY JOB SUBMISSIONS
 
@@ -131,7 +153,7 @@ class TestLoadSharingFacility(unittest.TestCase):
         for f in jobs + [array_script, array_jobs]:
             os.unlink(f)
 
-    def test_sub_2(self):
+    def test_sub_3(self):
         jobs = [make_script(["sleep 1"]) for _ in range(5)]
         array_script, array_jobs = prep_array_script(jobs, os.getcwd(), LoadSharingFacility.TASK_ENV)
         jobid = LoadSharingFacility.sub(array_script, array=[1, 5], hold=True, name=inspect.stack()[0][3])
@@ -141,7 +163,7 @@ class TestLoadSharingFacility(unittest.TestCase):
         for f in jobs + [array_script, array_jobs]:
             os.unlink(f)
 
-    def test_sub_3(self):
+    def test_sub_4(self):
         directory = os.getcwd()
         jobs = [make_script([["sleep 5"], ['echo "file {0}"'.format(i)]], directory=directory)
                 for i in range(5)]
@@ -162,11 +184,10 @@ class TestLoadSharingFacility(unittest.TestCase):
             for i, j in enumerate(jobs):
                 f = j.replace(".sh", ".log")
                 self.assertTrue(os.path.isfile(f))
-                content = open(f).read().strip()
-                self.assertEqual("file {0}".format(i), content)
+                self.assertEqual("file {0}".format(i), open(f).read().strip())
                 os.unlink(f)
 
-    def test_sub_4(self):
+    def test_sub_5(self):
         directory = os.getcwd()
         jobs = [make_script(['echo "file {0}"'.format(i)], directory=directory)
                 for i in range(100)]
@@ -188,11 +209,10 @@ class TestLoadSharingFacility(unittest.TestCase):
             for i, j in enumerate(jobs):
                 f = j.replace(".sh", ".log")
                 self.assertTrue(os.path.isfile(f))
-                content = open(f).read().strip()
-                self.assertEqual("file {0}".format(i), content)
+                self.assertEqual("file {0}".format(i), open(f).read().strip())
                 os.unlink(f)
 
-    def test_sub_5(self):
+    def test_sub_6(self):
         jobs = [make_script(["echo $LSF_BINDIR"], directory=os.getcwd()) for _ in range(2)]
         array_script, array_jobs = prep_array_script(jobs, os.getcwd(), LoadSharingFacility.TASK_ENV)
         jobid = LoadSharingFacility.sub(array_script, array=[1, 2], name=inspect.stack()[0][3])
@@ -211,8 +231,31 @@ class TestLoadSharingFacility(unittest.TestCase):
             for i, j in enumerate(jobs):
                 f = j.replace(".sh", ".log")
                 self.assertTrue(os.path.isfile(f))
-                content = open(f).read().strip()
-                self.assertEqual(os.environ["LSF_BINDIR"], content)
+                self.assertEqual(os.environ["LSF_BINDIR"], open(f).read().strip())
+                os.unlink(f)
+
+    def test_sub_7(self):
+        assert "PYJOB_ENV1" not in os.environ
+        os.environ["PYJOB_ENV1"] = "pyjob_random1"
+        jobs = [make_script(["echo $PYJOB_ENV1"], directory=os.getcwd()) for _ in range(2)]
+        jobid = LoadSharingFacility.sub(jobs, directory=os.getcwd(), name=inspect.stack()[0][3])
+        time.sleep(5)
+        start, timeout = time.time(), False
+        while LoadSharingFacility.stat(jobid):
+            # Don't wait too long, one minute, then fail
+            if ((time.time() - start) // 60) >= 1:
+                LoadSharingFacility.kill(jobid)
+                timeout = True
+            time.sleep(10)
+        for f in jobs:
+            os.unlink(f)
+        if timeout:
+            self.assertEqual(1, 0, "Timeout")
+        else:
+            for i, j in enumerate(jobs):
+                f = j.replace(".sh", ".log")
+                self.assertTrue(os.path.isfile(f))
+                self.assertEqual(os.environ["PYJOB_ENV1"], open(f).read().strip())
                 os.unlink(f)
